Feature/(frontend)bind data between frontend and backend #51

+++ backend/internal/routes/backup/backup.services.go
@@ -7,6 +7,8 @@ import (
 	"fmt"
 	"github.com/gin-gonic/gin"
 	"net/http"
+	"os"
+	"path/filepath"
 )
 
 var backupRepository = new(BackUpRepository)
@@ -75,15 +77,30 @@ func (b *BackupService) downloadBackUpFile(c *gin.Context) {
 		utils.HandleHTTPError(c, err.Error(), "Can not retrieve backup", http.StatusBadRequest)
 		return
 	}
-	// Format CreatedAt timestamp as part of the filename
-	formattedDate := backup.CreatedAt.Format("20060102") // YYYYMMDD format
-	filename := fmt.Sprintf("backup_%s.sql", formattedDate)
 
-	c.FileAttachment(backup.Filename, filename)
+	// Format CreatedAt timestamp as part of the filename
+	path := backup.Filename
+	filename := fmt.Sprintf("backup_%s.sql", filepath.Base(path))
+	fileInfo, _ := os.Stat(path)
+	// Set the appropriate headers
+	c.Header("Access-Control-Expose-Headers", "Content-Disposition")
+	c.Header("Content-Description", "File Transfer")
+	c.Header("Content-Disposition", fmt.Sprintf("attachment; filename=%s", filename))
+	c.Header("Content-Type", "application/octet-stream")
+	c.Header("Content-Transfer-Encoding", "binary")
+	c.Header("Content-Length", fmt.Sprintf("%d", fileInfo.Size()))
+
+	// Send the file
+	c.File(path)
 }
 
 func (b *BackupService) deleteBackUp(c *gin.Context) {
 	id := c.Param("id")
+	if id == "" {
+		utils.HandleHTTPError(c, "ID is required", "Can not delete backup", http.StatusBadRequest)
+		return
+	}
+
 	//Delete the backup record from database
 	filename, err := backupRepository.Delete(id)
 	if err != nil {
